+++ frontend/src/features/dashboard/types.ts
@@ -1,0 +1,1 @@
+export type TelemetryMessage = Record<string, unknown> | null;
