+++ frontend/src/lib/api/httpClient.ts
@@ -1,0 +1,19 @@
+const API_BASE_URL = import.meta.env.VITE_API_BASE_URL ?? '';
+
+export const httpRequest = async <T>(path: string, init?: RequestInit): Promise<T> => {
+  const response = await fetch(`${API_BASE_URL}${path}`, {
+    ...init,
+    headers: {
+      'Content-Type': 'application/json',
+      ...(init?.headers ?? {}),
+    },
+  });
+
+  if (!response.ok) {
+    throw new Error(`Request failed: ${response.status}`);
+  }
+
+  return response.json() as Promise<T>;
+};
+
+export const getApiBaseUrl = () => API_BASE_URL;
